Moved battery state to binary sensor platform.
