+++ homeassistant/components/bosch_shc/binary_sensor.py
@@ -1,7 +1,10 @@
 """Platform for binarysensor integration."""
-from boschshcpy import SHCSession, SHCShutterContact
+import logging
+
+from boschshcpy import SHCBatteryDevice, SHCSession, SHCShutterContact
 
 from homeassistant.components.binary_sensor import (
+    DEVICE_CLASS_BATTERY,
     DEVICE_CLASS_DOOR,
     DEVICE_CLASS_WINDOW,
     BinarySensorEntity,
@@ -9,6 +12,8 @@ from homeassistant.components.binary_sensor import (
 
 from .const import DATA_SESSION, DOMAIN
 from .entity import SHCEntity
+
+_LOGGER = logging.getLogger(__name__)
 
 
 async def async_setup_entry(hass, config_entry, async_add_entities):
@@ -24,6 +29,25 @@ async def async_setup_entry(hass, config_entry, async_add_entities):
                 entry_id=config_entry.entry_id,
             )
         )
+
+    for binary_sensor in (
+        session.device_helper.motion_detectors
+        + session.device_helper.shutter_contacts
+        + session.device_helper.smoke_detectors
+        + session.device_helper.thermostats
+        + session.device_helper.twinguards
+        + session.device_helper.universal_switches
+        + session.device_helper.wallthermostats
+        + session.device_helper.water_leakage_detectors
+    ):
+        if binary_sensor.supports_batterylevel:
+            entities.append(
+                BatterySensor(
+                    device=binary_sensor,
+                    parent_id=session.information.unique_id,
+                    entry_id=config_entry.entry_id,
+                )
+            )
 
     if entities:
         async_add_entities(entities)
@@ -47,3 +71,47 @@ class ShutterContactSensor(SHCEntity, BinarySensorEntity):
             "GENERIC": DEVICE_CLASS_WINDOW,
         }
         return switcher.get(self._device.device_class, DEVICE_CLASS_WINDOW)
+
+
+class BatterySensor(SHCEntity, BinarySensorEntity):
+    """Representation of a SHC battery reporting sensor."""
+
+    @property
+    def unique_id(self):
+        """Return the unique ID of this sensor."""
+        return f"{self._device.serial}_battery"
+
+    @property
+    def name(self):
+        """Return the name of this sensor."""
+        return f"{self._device.name} Battery"
+
+    @property
+    def is_on(self):
+        """Return the state of the sensor."""
+        if (
+            self._device.batterylevel
+            == SHCBatteryDevice.BatteryLevelService.State.NOT_AVAILABLE
+        ):
+            _LOGGER.debug("Battery state of device %s is not available", self.name)
+
+        if (
+            self._device.batterylevel
+            == SHCBatteryDevice.BatteryLevelService.State.CRITICAL_LOW
+        ):
+            _LOGGER.warning("Battery state of device %s is critical low", self.name)
+
+        if (
+            self._device.batterylevel
+            == SHCBatteryDevice.BatteryLevelService.State.LOW_BATTERY
+        ):
+            _LOGGER.warning("Battery state of device %s is low", self.name)
+
+        return (
+            self._device.batterylevel != SHCBatteryDevice.BatteryLevelService.State.OK
+        )
+
+    @property
+    def device_class(self):
+        """Return the class of the sensor."""
+        return DEVICE_CLASS_BATTERY
